refactor: split profileStore and profile API to prevent bloat, fix isLoading race condition
Split monolithic profileStore into 3 focused stores (profileStore, pictureStore, tagStore)
each with independent isLoading/error state, fixing the concurrent fetch race condition.
Split api/profile.ts into api/profile.ts, api/pictures.ts, api/tags.ts.
Updated all consumers and tests.
