+++ web/src/features/profile/pages/EditProfilePage.tsx
@@ -1,5 +1,7 @@
 import { useEffect, type FC } from 'react';
 import { useProfileStore } from '@/stores/profileStore';
+import { usePictureStore } from '@/stores/pictureStore';
+import { useTagStore } from '@/stores/tagStore';
 import { ProfileForm } from '@/features/profile/components/ProfileForm';
 import { PhotoUploader } from '@/features/profile/components/PhotoUploader';
 import { TagManager } from '@/features/profile/components/TagManager';
@@ -8,18 +10,24 @@ import type { ProfileFormData } from '@/lib/validators';
 
 const EditProfilePage: FC = () => {
   const profile = useProfileStore((s) => s.profile);
-  const pictures = useProfileStore((s) => s.pictures);
-  const tags = useProfileStore((s) => s.tags);
-  const allTags = useProfileStore((s) => s.allTags);
-  const isLoading = useProfileStore((s) => s.isLoading);
-  const error = useProfileStore((s) => s.error);
+  const isProfileLoading = useProfileStore((s) => s.isLoading);
+  const profileError = useProfileStore((s) => s.error);
   const saveProfile = useProfileStore((s) => s.saveProfile);
   const fetchProfile = useProfileStore((s) => s.fetchProfile);
-  const fetchTags = useProfileStore((s) => s.fetchTags);
-  const uploadPicture = useProfileStore((s) => s.uploadPicture);
-  const deletePicture = useProfileStore((s) => s.deletePicture);
-  const addTag = useProfileStore((s) => s.addTag);
-  const removeTag = useProfileStore((s) => s.removeTag);
+
+  const pictures = usePictureStore((s) => s.pictures);
+  const isPicturesLoading = usePictureStore((s) => s.isLoading);
+  const pictureError = usePictureStore((s) => s.error);
+  const uploadPicture = usePictureStore((s) => s.uploadPicture);
+  const deletePicture = usePictureStore((s) => s.deletePicture);
+
+  const tags = useTagStore((s) => s.tags);
+  const allTags = useTagStore((s) => s.allTags);
+  const isTagsLoading = useTagStore((s) => s.isLoading);
+  const tagError = useTagStore((s) => s.error);
+  const fetchTags = useTagStore((s) => s.fetchTags);
+  const addTag = useTagStore((s) => s.addTag);
+  const removeTag = useTagStore((s) => s.removeTag);
 
   useEffect(() => {
     fetchProfile();
@@ -39,7 +47,9 @@ const EditProfilePage: FC = () => {
     });
   };
 
-  if (isLoading && !profile) {
+  const error = profileError ?? pictureError ?? tagError;
+
+  if (isProfileLoading && !profile) {
     return (
       <div className="flex items-center justify-center p-12">
         <Spinner size="lg" />
@@ -59,7 +69,7 @@ const EditProfilePage: FC = () => {
 
       <ProfileForm
         onSubmit={handleSubmit}
-        isLoading={isLoading}
+        isLoading={isProfileLoading}
         initialValues={profile ? {
           firstName: profile.firstName ?? undefined,
           lastName: profile.lastName ?? undefined,
@@ -76,7 +86,7 @@ const EditProfilePage: FC = () => {
         pictures={[...pictures]}
         onUpload={uploadPicture}
         onDelete={deletePicture}
-        isLoading={isLoading}
+        isLoading={isPicturesLoading}
       />
 
       <TagManager
@@ -84,7 +94,7 @@ const EditProfilePage: FC = () => {
         allTags={[...allTags]}
         onAdd={addTag}
         onRemove={removeTag}
-        isLoading={isLoading}
+        isLoading={isTagsLoading}
       />
     </div>
   );
