+++ web/src/tests/features/profile/pages/ProfileCreatePage.test.tsx
@@ -4,8 +4,12 @@ import userEvent from '@testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
 import { ProfileCreatePage } from '@/features/profile/pages/ProfileCreatePage';
 import { useProfileStore } from '@/stores/profileStore';
+import { usePictureStore } from '@/stores/pictureStore';
+import { useTagStore } from '@/stores/tagStore';
 
 vi.mock('@/stores/profileStore');
+vi.mock('@/stores/pictureStore');
+vi.mock('@/stores/tagStore');
 
 const mockSaveProfile = vi.fn();
 const mockFetchTags = vi.fn();
@@ -14,35 +18,62 @@ const mockDeletePicture = vi.fn();
 const mockAddTag = vi.fn();
 const mockRemoveTag = vi.fn();
 
-function setupMockStore(overrides: Record<string, unknown> = {}) {
+function setupMockStores(overrides: Record<string, unknown> = {}) {
+  const profileState = {
+    profile: null,
+    isLoading: false,
+    error: null,
+    saveProfile: mockSaveProfile,
+    fetchProfile: vi.fn(),
+    clearError: vi.fn(),
+    ...('error' in overrides ? { error: overrides.error } : {}),
+  };
+
+  const pictureState = {
+    pictures: [],
+    isLoading: false,
+    error: null,
+    uploadPicture: mockUploadPicture,
+    deletePicture: mockDeletePicture,
+    clearError: vi.fn(),
+  };
+
+  const tagState = {
+    tags: [],
+    allTags: [{ id: 1, name: 'hiking' }, { id: 2, name: 'cooking' }],
+    isLoading: false,
+    error: null,
+    fetchTags: mockFetchTags,
+    addTag: mockAddTag,
+    removeTag: mockRemoveTag,
+    clearError: vi.fn(),
+  };
+
   vi.mocked(useProfileStore).mockImplementation((selector: unknown) => {
-    const state = {
-      profile: null,
-      pictures: [],
-      tags: [],
-      allTags: [{ id: 1, name: 'hiking' }, { id: 2, name: 'cooking' }],
-      isLoading: false,
-      error: null,
-      saveProfile: mockSaveProfile,
-      fetchTags: mockFetchTags,
-      uploadPicture: mockUploadPicture,
-      deletePicture: mockDeletePicture,
-      addTag: mockAddTag,
-      removeTag: mockRemoveTag,
-      clearError: vi.fn(),
-      fetchProfile: vi.fn(),
-      ...overrides,
-    };
     if (typeof selector === 'function') {
-      return (selector as (s: typeof state) => unknown)(state);
+      return (selector as (s: typeof profileState) => unknown)(profileState);
     }
-    return state;
+    return profileState;
+  });
+
+  vi.mocked(usePictureStore).mockImplementation((selector: unknown) => {
+    if (typeof selector === 'function') {
+      return (selector as (s: typeof pictureState) => unknown)(pictureState);
+    }
+    return pictureState;
+  });
+
+  vi.mocked(useTagStore).mockImplementation((selector: unknown) => {
+    if (typeof selector === 'function') {
+      return (selector as (s: typeof tagState) => unknown)(tagState);
+    }
+    return tagState;
   });
 }
 
 beforeEach(() => {
   vi.resetAllMocks();
-  setupMockStore();
+  setupMockStores();
 });
 
 function renderPage() {
@@ -112,7 +143,7 @@ describe('ProfileCreatePage', () => {
   });
 
   it('displays error message when error is set', () => {
-    setupMockStore({ error: 'Something went wrong' });
+    setupMockStores({ error: 'Something went wrong' });
     renderPage();
 
     expect(
